[수정] baseFiles 분리
