+++ package-kuring/Sources/Caches/Dependency/UserDefault.swift
@@ -1,0 +1,30 @@
+//
+// Copyright (c) 2024 쿠링
+// See the 'License.txt' file for licensing information.
+//
+
+import Foundation
+
+@propertyWrapper
+struct UserDefault<T: Codable> {
+    let key: String
+    let defaultValue: T
+    let storage: UserDefaults
+
+    var wrappedValue: T {
+        get {
+            guard let data = self.storage.object(forKey: self.key) as? Data else { return defaultValue }
+            return (try? PropertyListDecoder().decode(T.self, from: data)) ?? self.defaultValue
+        }
+        set {
+            let encodedData = try? PropertyListEncoder().encode(newValue)
+            self.storage.set(encodedData, forKey: self.key)
+        }
+    }
+
+    init(key: String, defaultValue: T, storage: UserDefaults = .standard) {
+        self.key = key
+        self.defaultValue = defaultValue
+        self.storage = storage
+    }
+}
